Adds the structure improvements to the html version
